FS-91: round the expectation the way the GPU rounded the reality, and the rebase zero is hard again
The physics lane found that `FactoryStale.ts` differences an f64 composition
against the float32 `matrixAt` readback with no `Math.fround` anywhere in the
file. Confirmed at HEAD, and their fix is better than mine.

FS-89 measured the residual, refused to widen a threshold until it passed, and
DERIVED a floor instead: four float32 ULP at the engine magnitude, 9.537e-5 m at
a 200 m rebase threshold. That was principled and it was still an accommodation.
`Math.fround` narrows the expectation through the same single-precision step the
stored matrix already went through, so the two agree BIT FOR BIT and there is
nothing left to tolerate.

WHY A HARD ZERO IS WORTH MORE THAN A CORRECT TOLERANCE, which is the part worth
recording. The ULP floor GREW WITH THE THRESHOLD: 9.5e-5 m at 200 m, but 1.9e-3 m
at the shipped 4 km. It was never wrong at either. It was a number that had to be
recomputed and re-argued every time the threshold moved, which is exactly how a
bound eventually gets loosened by somebody who does not know why it is that size.
A hard zero cannot rot, and it cannot be widened by half a decimal without the
change being obvious.

They hit the identical thing on a launch pad, at 1e-6 m at 137 m, 4e-6 at 244 m
and 6e-6 at 367 m, all exactly 2^-23 times the distance, and correctly refused to
call it a slow leak. Same arithmetic, two lanes, one file each.

MEASURED, driven, walked and never teleported:

  200 m threshold, 2 rebases     worst staleMaxM 0            exactly
  1000 m threshold, 2 rebases    worst staleMaxM 0            exactly
  before the walk                staleMaxM 0 over 2 drawn
  body-frame movement            0 m
  drawnParts                     2, unchanged

The second row is the claim the fix actually makes: the zero holds at five times
the magnitude, where the old floor would have had to be five times looser.

NEGATIVE CONTROL, re-run because the assertion CHANGED from `< bound` to
`=== 0` and a control that was only ever run against the old form proves nothing
about the new one:

  FAIL nothingWentStaleDuringTheWalk  3.9943e+2 m, bound 0 m exactly, which is
       2.00e+0 of one rebase delta against 1.0 for a detachment
  FAIL nothingIsStaleNow              3.9943e+2 m over 2 drawn

with `somethingIsDrawn`, `theWalkCrossedARebase`, `stillDrawn`,
`notStaleBeforeTheWalk` and `thePlanDidNotMove` all still passing. Applied in the
scratch tree only, so the shared checkout never held it; verified after with
`ctlPlaced` 0 in both trees, `diff -q` byte-identical, and `git status
--porcelain` on `FactoryView.ts` printing nothing.

`staleFraction` is kept, because it is scale free and it is what makes the
separation legible: a detached instance reads 1.0, and the control reads 2.00
because two rebase deltas accumulate on an instance that is placed once.

THE SECOND REPORTED FINDING DOES NOT REPRODUCE AND IS RECORDED AS NOT A DEFECT.
`probes/factoryrebase.js` was reported to call `mustNum(of.world(), 'rebaseM')`
at line 66 against a field that lives on `Config`. At HEAD line 87 reads
`OF_ARGS.rebaseM ?? null` and line 61 is a comment correctly naming
`Config.rebaseM`; there is no such call and the probe's own results demonstrate
it runs. Recorded rather than dropped, because the fourth cross-context claim
today needing verification at the source is a process fact and not bad luck, and
because "an instrument's stated blind spot is the list of claims that cannot have
come from it" applies to relayed diagnoses as well as to tools.

+++ web/src/game/FactoryStale.ts
@@ -63,7 +63,28 @@ export function stalenessOf(f: Factory | null,
     if (m === null) continue;
     out.drawn++;
     origin.toEngine(p.pos, scratch);
-    const d = Math.hypot(m[12] - scratch.x, m[13] - scratch.y, m[14] - scratch.z);
+    // FS-91: ROUND THE EXPECTATION THE WAY THE GPU ROUNDED THE REALITY.
+    //
+    // `matrixAt` reads back out of a `BatchedMesh`, which stores instance
+    // matrices as FLOAT32 because they are on their way to the GPU, and
+    // `toEngine` composes in f64. Differencing the two directly measures the
+    // narrowing as well as the staleness, and FS-89 accommodated that with a
+    // bound derived from four float32 ULP at the engine magnitude.
+    //
+    // THE ACCOMMODATION WAS PRINCIPLED AND THIS IS BETTER, and it came from the
+    // physics lane hitting the identical thing on a launch pad: 1e-6 m at 137 m,
+    // 4e-6 at 244 m and 6e-6 at 367 m, which is 2^-23 times the distance and is
+    // therefore not a leak. `Math.fround` narrows the expectation through the
+    // same single-precision step the stored matrix already went through, so the
+    // two agree BIT FOR BIT and the correct answer is a hard 0 again.
+    //
+    // A HARD ZERO THAT STAYS HARD AT ANY DISTANCE IS WORTH MORE THAN A DERIVED
+    // TOLERANCE, and the reason is not tidiness. The ULP floor grew with the
+    // rebase threshold, so at the shipped 4 km it was 1.9e-3 m: still seven
+    // orders below a detachment, and still a number that has to be recomputed
+    // and re-argued every time the threshold moves. This one never moves.
+    const d = Math.hypot(m[12] - Math.fround(scratch.x),
+      m[13] - Math.fround(scratch.y), m[14] - Math.fround(scratch.z));
     if (d > 0) out.stale++;
     if (d > out.maxM) { out.maxM = d; out.worst = p.id; }
   }
